Simplify configuration
