+++ code/core/src/core-server/presets/vitePlugins/vite-inject-mocker/constants.ts
@@ -1,2 +1,1 @@
-export const VIRTUAL_MODULE_MOCKER_INTERCEPTOR = 'virtual:module-mocker-interceptor';
 export const __STORYBOOK_GLOBAL_THIS_ACCESSOR__ = '__vitest_mocker__';
